fix: database is locked on multiple offline statuses
https://sqlite.org/lang_transaction.html

This is an attempt to solve a flaky but common "database is locked" error
when two statuses record offline at the same time.

The record_offline() function starts with SELECT FROM status_history,
followed by INSERT INTO status_history, which upgrades the read transaction
to a write transaction.

I'm not fully sure of this, but my guess is that when two statuses are
offline, there's a chance that they'll both open a transaction and begin
a read which, under WAL mode, would be allowed to run concurrently.
One of the transactions runs the INSERT statement first, upgrading to
a write transaction, which again can work concurrently with the second
read transaction. However, once the second transaction runs the same
INSERT statement, it wants to upgrade to a write transaction too, but
the first transaction is already writing, and SQLite prohibits concurrent
writers even in WAL mode, therefore raising the locked exception.

It seems that this upgrade can't be deferred by sqlite3's timeout parameter,
as this error can occur regardless of what timeout is used.

There are three solutions I can think of:
1. Extract the SELECT to a separate connection
2. Turn off the transaction and rely on autocommit
3. Start a write transaction immediately

(1) is easy to do, but additional connections incur a fair performance hit.
(2) is easy to do, but breaks atomicity with SELECT + INSERT.
(3) requires tweaking connect(), but retains everything in a single transaction.

This commit attempts solution (3).